create new rout ,admin file ,some fix

+++ TodoApp/routers/todos.py
@@ -36,12 +36,16 @@ class TodoRequest(BaseModel):
 # get all todos
 @router.get("/", status_code=status.HTTP_200_OK)
 async def read_all(user:user_dependency,db: db_dependency):
+    if user is None:
+        raise HTTPException(status_code=401,detail ="Authentications Failed")
     return db.query(Todo).filter(Todo.owner_id==user.get('id')).all()
 
 # get todo by id using path parameter
 @router.get("/todo/{todo_id}", status_code=status.HTTP_200_OK)
-async def read_todo(db: db_dependency, todo_id: int = Path(gt=0)):
-    todo_model = db.query(Todo).filter(Todo.id == todo_id).first()
+async def read_todo(user:user_dependency,db: db_dependency, todo_id: int = Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=401,detail ="Authentications Failed")
+    todo_model = db.query(Todo).filter(Todo.id == todo_id).filter(Todo.owner_id == user.get('id')).first()
     if todo_model is not None:
         return todo_model
     raise HTTPException(status_code=404, detail="Todo not found")
@@ -59,12 +63,15 @@ async def create_todo(user:user_dependency,db: db_dependency, todo_request: Todo
 # update todo by id using put method
 @router.put("/todo/{todo_id}", status_code=status.HTTP_200_OK)
 async def update_todo(
+    user:user_dependency,
     db: db_dependency,
     todo_request: TodoRequest,
     todo_id: int = Path(gt=0)
     
 ):
-    todo_model = db.query(Todo).filter(Todo.id == todo_id).first()
+    if user is None:
+        raise HTTPException(status_code=401,detail ="Authentications Failed")
+    todo_model = db.query(Todo).filter(Todo.id == todo_id).filter(Todo.owner_id == user.get('id')).first()
 
     if todo_model is None:
         raise HTTPException(status_code=404, detail="Todo not found")
@@ -80,11 +87,13 @@ async def update_todo(
 
 # delete todo by id using delete method
 @router.delete("/todo/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_todo(db: db_dependency, todo_id: int = Path(gt=0)):
-    todo_model = db.query(Todo).filter(Todo.id == todo_id).first()
+async def delete_todo(user:user_dependency,db: db_dependency, todo_id: int = Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=401,detail ="Authentications Failed")
+    todo_model = db.query(Todo).filter(Todo.id == todo_id).filter(Todo.owner_id == user.get('id')).first()
 
     if todo_model is None:
         raise HTTPException(status_code=404, detail="Todo not found")
-    db.query(Todo).filter(Todo.id == todo_id).delete()
+    db.query(Todo).filter(Todo.id == todo_id).filter(Todo.owner_id == user.get('id')).delete()
 
     db.commit()
